Revert "Update Dienblad.java"
This reverts commit c52574fce295f3c3ab4d920a1565c2b1ebd482b6.

+++ src/main/java/Dienblad.java
@@ -55,7 +55,7 @@ public class Dienblad {
         Iterator<Artikel> it = artikelen.iterator();
         while(it.hasNext()) {
             Artikel artikel = it.next();
-            totaal += artikel.getPrijs().intValue();
+            totaal += artikel.getPrijs();
         }
         return totaal;
     }
